incluidos testes de validação de delete em author e book

+++ backend/src/main/service/author_service.py
@@ -1,7 +1,8 @@
+import sqlalchemy
 from backend.src.main import db
 from main.model.author import Author, AuthorSchema
 from main.model.series import Series
-from main.util.utils import response_created, response_conflict, response_success
+from main.util.utils import response_created, response_conflict, response_success, response_bad_request
 
 
 def upsert_author(data, update):
@@ -16,6 +17,7 @@ def upsert_author(data, update):
             db.session.commit()
             return response_success('Author successfully updated.')
         else:
+
             author = AuthorSchema().load(data).data
             author.series = series
             save_changes(author)
@@ -33,9 +35,20 @@ def get_an_author(author_id):
 
 
 def delete_author(author_id):
-    Author.query.filter_by(id=author_id).delete()
-    db.session.commit()
-    return None, 201
+    author = Author.query.get(author_id)
+    if author:
+        if has_no_dependencies(author):
+            Author.query.filter_by(id=author_id).delete()
+            db.session.commit()
+            return response_success('')
+        else:
+            return response_conflict("Author has dependencies and can't be deleted.")
+    else:
+        return response_bad_request("Author not found.")
+
+
+def has_no_dependencies(author):
+    return not author.books and not author.series
 
 
 def save_changes(data):
